ApplicationGestionAi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
mysql-connector-python==8.2.0
pydantic==2.5.0
python-multipart==0.0.6
